FastAPI-Server/security/auth.py
# security/auth.py

from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt, jwe
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer

from config.settings import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/auth/login")

# Генерация JWT токена
def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire, "type": "access"})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.JWT_ALGORITHM)

# Генерация JWE refresh-токена (зашифрован)
def create_refresh_token(data: dict) -> str:
    to_encode = data.copy()
    to_encode.update({"type": "refresh"})
    # Шифруем токен
    return jwe.encrypt(
        jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.JWT_ALGORITHM),
        settings.SECRET_KEY.encode(),
        algorithm="dir",
        encryption="A256GCM"
    ).decode('utf-8')

# Расшифровка и проверка refresh-токена
def verify_refresh_token(token: str) -> dict:
    try:
        decrypted = jwe.decrypt(token.encode(), settings.SECRET_KEY.encode())
        payload = jwt.decode(decrypted, settings.SECRET_KEY, algorithms=[settings.JWT_ALGORITHM])
        if payload.get("type") != "refresh":
            raise JWTError("Invalid token type")
        return payload
    except Exception:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid refresh token",
            headers={"WWW-Authenticate": "Bearer"},
        )

# Хеширование пароля
def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

# Проверка пароля
def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

# Получение текущего пользователя (зависимость)
async def get_current_user(token: str = Depends(oauth2_scheme)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.JWT_ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        # Здесь можно добавить проверку в БД или кэш
        return {"user_id": user_id}
    except JWTError:
        raise credentials_exception