FastAPI-Server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.4
backports-datetime-fromisoformat==2.0.3
bandit==1.8.6
black==25.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
distlib==0.4.0
dodgy==0.2.1
dparse==0.6.4
exceptiongroup==1.3.0
fastapi==0.116.2
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mdurl==0.1.2
mypy==1.18.1
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psutil==7.1.0
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pytest==8.4.2
PyYAML==6.0.2
radon==6.0.1
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.0
safety==3.6.1
safety-schemas==0.0.14
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
stevedore==5.5.0
tenacity==9.1.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
vulture==2.14
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
Authlib==1.6.4
backports-datetime-fromisoformat==2.0.3
bandit==1.8.6
bcrypt==4.3.0
black==25.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
Deprecated==1.2.18
distlib==0.4.0
dodgy==0.2.1
dparse==0.6.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.2
filelock==3.19.1
ghp-import==2.1.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
limits==5.5.0
mando==0.7.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-get-deps==0.2.0
mkdocstrings==0.30.0
mkdocstrings-python==1.18.2
mypy==1.18.1
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
psutil==7.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
Pygments==2.19.2
pymdown-extensions==10.16.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
pyyaml_env_tag==1.1
radon==6.0.1
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.0
safety==3.6.1
safety-schemas==0.0.14
secure==1.0.1
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
stevedore==5.5.0
structlog==25.4.0
tenacity==9.1.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
vulture==2.14
watchdog==6.0.0
wrapt==1.17.3
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
Authlib==1.6.4
backports-datetime-fromisoformat==2.0.3
bandit==1.8.6
bcrypt==4.3.0
black==25.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.6
cryptography==46.0.1
Deprecated==1.2.18
distlib==0.4.0
dodgy==0.2.1
dparse==0.6.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.2
filelock==3.19.1
ghp-import==2.1.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
limits==5.5.0
mando==0.7.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-get-deps==0.2.0
mkdocstrings==0.30.0
mkdocstrings-python==1.18.2
mypy==1.18.1
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
psutil==7.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
Pygments==2.19.2
pymdown-extensions==10.16.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
pyyaml_env_tag==1.1
radon==6.0.1
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.0
safety==3.6.1
safety-schemas==0.0.14
secure==1.0.1
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
stevedore==5.5.0
structlog==25.4.0
tenacity==9.1.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
vulture==2.14
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
